% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascension_thursday.R
\name{ascension_thursday}
\alias{ascension_thursday}
\title{Create dummies for Ascension Day (In dutch: 'Hemelvaartdag')}
\usage{
ascension_thursday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Ascension Day. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after Ascension Day, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-5-24", "2021-5-31"),
ascension_day = ascension_thursday(dates)
)
print(df)
}
