% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_friday.R
\name{black_friday}
\alias{black_friday}
\title{Create dummies for Black Friday}
\usage{
black_friday(dates, pre = 0, post = 0, country = "US")
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}

\item{country}{"US" or "Canada"}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Black Friday. The day(s) before or after Black Friday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the day after Black Friday, set the
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2022-11-22", "2022-11-29"),
blck_fday = black_friday(dates, country="US")
)
df
}
