% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_holidays.R
\name{boxing_day}
\alias{boxing_day}
\title{Create dummies for boxing day (In dutch: 'Tweede kerstdag')}
\usage{
boxing_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{nonnegative integer. To encode the days before the holiday dates}

\item{post}{nonngevative integer. To encode the days after the holiday dates}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of boxing day. The day(s) before or after boxing day
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the two days after boxing day, set
post argument to 2.
}
\examples{
dates <- seq_days("2021-12-23", "2021-12-31")
tibble::tibble(
   dates,
   box_day = boxing_day(dates, pre = 1, post = 2)
)
}
