% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnival_sunday.R
\name{carnival_sunday}
\alias{carnival_sunday}
\title{Create dummies for Carnival Sunday (In dutch: 'Carnaval')}
\usage{
carnival_sunday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Carnival Sunday. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the day after Carnival Sunday, set the
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2022-1-1", "2022-12-31"),
carnival = carnival_sunday(dates, pre=0,post=0)
)
subset(df, carnival == 1)
}
