% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummies.R
\name{create_dummy}
\alias{create_dummy}
\title{Create a dummy variable from a date vector}
\usage{
create_dummy(dates, holiday_dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{date}

\item{holiday_dates}{date}

\item{pre}{non-negative int. To recode the x days after the holiday dates.}

\item{post}{non-negative int. To recode the x days before the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of a holiday. The day(s) before or after the holiday dates
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the two days after your specified holiday dates, set
post argument to 2.
}
\examples{
dates <- seq_days(from = "2021-12-24", to = "2021-12-31")
create_dummy(dates, as.Date("2021-12-25"), post = 1)
}
