% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DutchDayDummies.R
\docType{package}
\name{DutchDayDummies}
\alias{DutchDayDummies}
\title{DutchDayDummies: A package for computing regression dummies for dutch holidays.}
\description{
The DutchDayDummies package provides three categories of important functions:
official holidays, non-official holidays and vacations.
}
\section{Official holiday functions}{

The foo functions
}

