% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easter_sunday.R
\name{easter_sunday}
\alias{easter_sunday}
\title{Create dummies for Easter sunday (In dutch: 'Eerste Paasdag')}
\usage{
easter_sunday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Easter day. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after Easter sunday, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-4-5", "2023-4-10"),
easter_sun = easter_sunday(dates, post = 1)
)
subset(df, easter_sun == 1)
}
