% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_easter_algorithm.R
\name{gauss_easter_algorithm}
\alias{gauss_easter_algorithm}
\title{Gauss Easter Algorithm to calculate Easter dates}
\usage{
gauss_easter_algorithm(years)
}
\arguments{
\item{years}{int.}
}
\description{
Gauss Easter Algorithm to calculate Easter dates
}
\examples{
gauss_easter_algorithm(2021)
}
