% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekday_before_dates.R
\name{get_wday_before_dates}
\alias{get_wday_before_dates}
\title{Get the dates of a weekday before a specified date}
\usage{
get_wday_before_dates(weekday, dates, search = "before")
}
\arguments{
\item{weekday}{Character. Name of the week (e.g. "monday")}

\item{dates}{The date of interest. Format: YYYY-MM-DD.}

\item{search}{Search "before" of "after" the specified date.}
}
\value{
Character vector in the format YYYY-MM-DD.
}
\description{
Get the dates of a weekday before a specified date
}
\examples{
# search the monday before the "2021-12-29"
date_of_monday <- get_wday_before_dates(weekday = "monday", dates = "2021-12-29", search = "before")
# Confirm
weekdays(as.Date(date_of_monday))
}
