% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_friday.R
\name{good_friday}
\alias{good_friday}
\title{Create dummies for Good Friday (In dutch: 'Goede Vrijdag')}
\usage{
good_friday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-ngevative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Easter day. The day(s) before or after Good Friday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after Good Friday, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-4-1", "2021-4-10"),
good_fday = good_friday(dates)
)
print(df)
}
