% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_holidays.R
\name{holiday_vacations}
\alias{holiday_vacations}
\title{Load the dummies from all dutch vacations into a dataframe.}
\usage{
holiday_vacations(dates)
}
\arguments{
\item{dates}{date vector of format YYYY-MM-DD.}
}
\value{
data.frame of class tibble
}
\description{
The following vacations are included: \cr
\code{\link[DutchDayDummies]{autumn_vacation}} \cr
\code{\link[DutchDayDummies]{xmas_vacation}} \cr
\code{\link[DutchDayDummies]{spring_vacation}} \cr
\code{\link[DutchDayDummies]{may_vacation}} \cr
\code{\link[DutchDayDummies]{summer_vacation}} \cr
}
\examples{
dates <- lubridate::today()
df <- holiday_vacations(dates)
df
}
