% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_holidays.R
\name{holiday_weekends}
\alias{holiday_weekends}
\title{Load the data from all holiday weekends.}
\usage{
holiday_weekends(dates)
}
\arguments{
\item{dates}{date vector of format YYYY-MM-DD.}
}
\value{
data.frame of class tibble
}
\description{
The following holiday weekends are included: \cr
\code{\link[DutchDayDummies]{xmas_and_boxing_day}} \cr
\code{\link[DutchDayDummies]{pentecost_days}} \cr
\code{\link[DutchDayDummies]{easter_weekend}} \cr
}
\examples{
dates <- lubridate::today()
df <- holiday_weekends(dates)
df
}
