% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kings_day.R
\name{kings_day}
\alias{kings_day}
\title{Create dummies for king's day (In dutch: 'Koningsdag')}
\usage{
kings_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of king's day. The day(s) before or after king's day
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after king's day, set
post argument to 1.
}
\examples{
dates <- seq_days("2021-4-25", "2021-4-30")
tibble::tibble(
   dates,
   kday = kings_day(dates)
)
}
