% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_holidays.R
\name{load_all_days}
\alias{load_all_days}
\title{Return the dummies for each available holiday, official and non-officials, weekends and vacations.}
\usage{
load_all_days(dates)
}
\arguments{
\item{dates}{date vector of format YYYY-MM-DD.}
}
\value{
data.frame of class tibble
}
\description{
Return the dummies for each available holiday, official and non-officials, weekends and vacations.
}
\examples{
dates <- lubridate::today()
df <- load_all_days(dates)
df
}
