% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/may_vacation.R
\name{may_vacation}
\alias{may_vacation}
\title{Create dummies for May vacation (In dutch: 'Meivakantie')}
\usage{
may_vacation(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-ngevative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of may vacation. The day(s) before or after May vacation
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after May vacation, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-1-1", "2021-12-31"),
may_holiday = may_vacation(dates)
)
#View(df)
}
