% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyear_holiday.R
\name{newyears_day}
\alias{newyears_day}
\title{Create dummies for new year's day (In dutch: 'Nieuwjaarsdag')}
\usage{
newyears_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{nonnegative integer. To encode the days before the holiday dates}

\item{post}{nonngevative integer. To encode the days after the holiday dates}
}
\value{

}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of new year's day. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the two days after new year's day, set
post argument to 2.
}
\examples{
dates <- seq_days("2021-12-27", "2022-1-2")
tibble::tibble(
   dates,
   nyd = newyears_day(dates)
)
}
