% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyear_holiday.R
\name{newyears_eve}
\alias{newyears_eve}
\title{Create dummies for new year's eve (In dutch: 'Oudjaarsdag')}
\usage{
newyears_eve(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of new year's eve. The day(s) before or after new year's eve
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the two days after new year's eve, set
post argument to 2.
}
\examples{
dates <- seq_days("2021-12-27", "2022-1-2")
tibble::tibble(
   dates,
   nye = newyears_eve(dates)
)
}
