% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_holiday.R
\name{next_holiday}
\alias{next_holiday}
\title{Find the number of days until the next holiday.}
\usage{
next_holiday(date = lubridate::today())
}
\value{
list
}
\description{
Find the number of days until the next holiday.
}
\examples{
next_holiday()
}
