% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_holidays.R
\name{non_official_holidays}
\alias{non_official_holidays}
\title{Load the dummies from all non-official Holidays in a data frame.}
\usage{
non_official_holidays(dates)
}
\arguments{
\item{dates}{date vector of format YYYY-MM-DD.}
}
\value{
data.frame of class tibble
}
\description{
The following holidays are included: \cr
\code{\link[DutchDayDummies]{carnival_sunday}} \cr
\code{\link[DutchDayDummies]{valentines_day}} \cr
\code{\link[DutchDayDummies]{good_friday}} \cr
\code{\link[DutchDayDummies]{ramadan}} \cr
\code{\link[DutchDayDummies]{remembrance_day}} \cr
\code{\link[DutchDayDummies]{liberation_day}} \cr
\code{\link[DutchDayDummies]{sugar_feast}} \cr
\code{\link[DutchDayDummies]{fathers_day}} \cr
\code{\link[DutchDayDummies]{sacrificial_feast}} \cr
\code{\link[DutchDayDummies]{prinsjesdag}} \cr
\code{\link[DutchDayDummies]{world_animal_day}} \cr
\code{\link[DutchDayDummies]{thanksgiving_day}} \cr
\code{\link[DutchDayDummies]{halloween}} \cr
\code{\link[DutchDayDummies]{st_martins_day}} \cr
\code{\link[DutchDayDummies]{black_friday}} \cr
\code{\link[DutchDayDummies]{cyber_monday}} \cr
\code{\link[DutchDayDummies]{st_nicholas_day}} \cr
\code{\link[DutchDayDummies]{newyears_eve}} \cr
}
\examples{
dates <- lubridate::today()
df <- non_official_holidays(dates)
df
}
