% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_weekday_of_a_month.R
\name{nth_weekday_of_a_month}
\alias{nth_weekday_of_a_month}
\title{Function to find the nth weekday in a certain month}
\usage{
nth_weekday_of_a_month(dates, nth, weekday, month)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{nth}{int. The nth weekend you want to search for given a day and month}

\item{weekday}{int. Sunday counts as 1, monday as 2, etc.}

\item{month}{int. The month of interest (1 for January, 2 for February etc.)}
}
\value{
Date written in type character.
}
\description{
Function to find the nth weekday in a certain month
}
\examples{
dates <- seq_days(from = "2021-1-1", to = "2021-12-31")
# Suppose you want to find the date of mother's day
# Mother's day falls on the second sunday of may, thus:
nth_weekday_of_a_month(dates, nth = 2, weekday = 1, month = 5)
}
