% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_holidays.R
\name{official_holidays}
\alias{official_holidays}
\title{Load the dummy variables from all Dutch official Holidays in a data frame.}
\usage{
official_holidays(dates)
}
\arguments{
\item{dates}{date vector of format YYYY-MM-DD.}
}
\value{
data.frame of class tibble
}
\description{
The following holidays are included: \cr
\code{\link[DutchDayDummies]{newyears_day}} \cr
\code{\link[DutchDayDummies]{easter_sunday}} \cr
\code{\link[DutchDayDummies]{easter_monday}} \cr
\code{\link[DutchDayDummies]{kings_day}} \cr
\code{\link[DutchDayDummies]{white_monday}} \cr
\code{\link[DutchDayDummies]{white_sunday}} \cr
\code{\link[DutchDayDummies]{xmas_day}} \cr
\code{\link[DutchDayDummies]{boxing_day}} \cr
}
\examples{
date <- lubridate::today()
df <- official_holidays(date)
df
}
