% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_days.R
\name{payment_days}
\alias{payment_days}
\title{Create dummies for payment days (In dutch: 'Betaaldag')}
\usage{
payment_days(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of paydays. The day(s) before or after payment days
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after payday, set
post argument to 1.
}
\examples{
dates <- seq_days("2021-1-1", "2021-12-31")
df <- tibble::tibble(
dates,
payday=payment_days(dates)
)
subset(df, payday == 1)
}
