% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pentecost_days.R
\name{pentecost_days}
\alias{pentecost_days}
\title{Create dummies for both Pentecost days (In dutch: 'Pinksteren')}
\usage{
pentecost_days(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of White Monday & Sunday. The day(s) before or after these holidays
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after the Pentecost, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-5-20", "2021-5-29"),
pentecost = pentecost_days(dates)
)
print(df)
}
