% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinsjesdag.R
\name{prinsjesdag}
\alias{prinsjesdag}
\title{Create dummies for Prinsjesdag}
\usage{
prinsjesdag(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Prinsjesdag. The day(s) before or after Prinsjesdag
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after Prinsjesdag, set
post argument to 1.
}
\examples{
dates <- seq_days("2021-9-19", "2021-9-27")
tibble::tibble(
dates,
prnsdag=prinsjesdag(dates)
)
}
