% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remembrance_day.R
\name{remembrance_day}
\alias{remembrance_day}
\title{Create dummies for Remembrance day (In dutch: 'Dodenherdenking')}
\usage{
remembrance_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Remembrance day. The day(s) before or after Remembrance day
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after Remembrance day, set
post argument to 1.
}
\examples{
dates <- seq_days("2021-5-1", "2021-5-10")
tibble::tibble(
   dates,
   remem_day = remembrance_day(dates)
)
}
