% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nicholas_day.R
\name{st_nicholas_day}
\alias{st_nicholas_day}
\title{Create dummies for St. Nicholoas Day (In dutch: 'Sinterklaas')}
\usage{
st_nicholas_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of St. Nicholas Day. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the day after Sinterklaas, set the
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2022-12-1", "2022-12-10"),
sinterklaas = st_nicholas_day(dates, pre=0,post=0)
)
subset(df, sinterklaas == 1)
}
