% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thanksgiving_day.R
\name{thanksgiving_day}
\alias{thanksgiving_day}
\title{Create dummies for Thanksgiving}
\usage{
thanksgiving_day(dates, pre = 0, post = 0, country = "US")
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-negative int. To encode the days after the holiday dates.}

\item{country}{"US" or "Canada"}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of Thanksgiving day. The day(s) before or after this holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the day after Thanksgiving, set the
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2022-11-22", "2022-11-29"),
thxgivin = thanksgiving_day(dates, country="US")
)
df
}
