% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ummalqura_algorithm.R
\name{ummalqura_algorithm}
\alias{ummalqura_algorithm}
\title{Umm al-Qura algorithm to conver Gregorian Dates to Islamic Dates}
\usage{
ummalqura_algorithm(date)
}
\arguments{
\item{date}{Date in format YYYY-MM-DD.}
}
\value{
Character
}
\description{
Function converts gregorian to islamic daes. Yields only valid dates for the period from 1937 to 2077.
}
\examples{
date <- lubridate::today()
# convert to islamic date
ummalqura_algorithm(date)
}
