% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_monday.R
\name{white_monday}
\alias{white_monday}
\title{Create dummies for White Monday (In dutch: 'Tweede Pinksterdag')}
\usage{
white_monday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of White Monday. The day(s) before or after White Monday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after White Monday, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-5-20", "2021-5-29"),
w_monday = white_monday(dates)
)
print(df)
}
