% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_sunday.R
\name{white_sunday}
\alias{white_sunday}
\title{Create dummies for White Sunday (In dutch: 'Eerste Pinksterdag')}
\usage{
white_sunday(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of White Sunday. The day(s) before or after White Sunday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after White Sunday, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-5-20", "2021-5-29"),
w_sunday = white_sunday(dates)
)
print(df)
}
