% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_animal_day.R
\name{world_animal_day}
\alias{world_animal_day}
\title{Create dummies for World Animal Day (In dutch: 'Dierendag')}
\usage{
world_animal_day(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD.}

\item{pre}{non-negative integer. To encode the days before the holiday dates.}

\item{post}{non-negative integer. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector).
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of World Animal Day. The day(s) before or after World Animal Day
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after World Animal Day, set
post argument to 1.
}
\examples{
dates <- seq_days("2021-3-31", "2021-4-6")
tibble::tibble(
dates,
animal_day=world_animal_day(dates)
)
}
