% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas_vacation.R
\name{xmas_vacation}
\alias{xmas_vacation}
\title{Create dummies for Christmas vacation (In dutch: 'Kerstvakantie')}
\usage{
xmas_vacation(dates, pre = 0, post = 0)
}
\arguments{
\item{dates}{A vector with dates in the format YYYY-MM-DD}

\item{pre}{non-negative int. To encode the days before the holiday dates.}

\item{post}{non-ngevative int. To encode the days after the holiday dates.}
}
\value{
A dummy (binary vector)
}
\description{
Create a dummy variable that takes the value 0 or 1 to indicate the
abscence or presence of christmas vacation. The day(s) before or after christmas holiday
can also be recoded by using the pre and post arguments. For example, if you also
want to recode the days after christmas vacation, set
post argument to 1.
}
\examples{
df <- tibble::tibble(
dates = seq_days("2021-1-1", "2021-12-31"),
xmas_holiday = xmas_vacation(dates)
)
#View(df)
}
